+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.1.1] - Unreleased
 - Added TS definitions for screenfull.
+- Added facet back button in modal form.
 
 ## [0.1.0] - 2018-05-24
 - Initial release